+++ src/index.js
@@ -1,1 +1,2 @@
 export { QueryBuilderPlugin as default } from './QueryBuilder';
+export { DynamicTable } from './QueryBuilder/ResultViewer';